🐛Fix: minor bugs (#77)
* 🐛 Fix: early return if created by me is false on dnd

* ✨ Feat: rerender after invite accepted

+++ src/components/common/sidebar/DashboardListItem.tsx
@@ -36,6 +36,7 @@ const DashboardListItem = () => {
 
     if (destination.droppableId === 'bin') {
       const item = dashboardListData[source.index];
+      if (!item.createdByMe) return;
       await instance.delete(`/dashboards/${item.id}`);
       router.push('/my-dashboard');
       handleLoadDashboardList();
